# backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "my-ToDoList API",
    "description": "인증된 사용자가 개인별 Todo를 카테고리 단위로 관리하는 API. Category는 Admin이 전역으로 관리하고, 일반 사용자는 조회/선택만 가능하다.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "로컬 개발 서버" }
  ],
...
  "paths": {
...
    "/auth/signup": {
      "post": {
        "tags": ["Auth"],
        "summary": "회원가입",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SignupRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "회원가입 성공",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } }
          },
          "400": {
            "description": "이메일 형식 오류 또는 이메일 중복(목록 내 중복 체크만 수행, 실제 이메일 존재 여부는 검증하지 않음)",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "로그인 (Access Token은 응답 바디, Refresh Token은 httpOnly 쿠키)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "로그인 성공. Refresh Token은 Set-Cookie(httpOnly)로 전달되며 응답 바디에는 포함되지 않는다.",
            "headers": {
              "Set-Cookie": {
                "description": "refreshToken=...; HttpOnly; Path=/; Secure",
                "schema": { "type": "string" }
              }
            },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginResponse" } } }
          },
          "400": {
            "description": "요청 형식 오류",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "401": {
            "description": "이메일/비밀번호 불일치(계정 존재 여부는 노출하지 않음)",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "Access Token 재발급 (httpOnly 쿠키의 Refresh Token 사용)",
        "responses": {
          "200": {
            "description": "재발급 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": { "type": "string", "description": "새로 발급된 JWT Access Token" },
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Refresh Token이 없거나 유효하지 않음(재로그인 필요)",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access Token을 `Authorization: Bearer {accessToken}` 헤더로 전달. Refresh Token은 httpOnly 쿠키로만 전달되어 이 스킴에 포함되지 않는다."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": { "type": "string", "example": "인증이 필요합니다." },
              "status": { "type": "integer", "example": 401 }
            },
            "required": ["message", "status"]
          }
        },
        "required": ["error"]
      },
      "Role": {
        "type": "string",
        "enum": ["Member", "Admin"]
      },
...
      "User": {
        "type": "object",
        "properties": {
          "userId": { "type": "integer", "example": 1 },
          "email": { "type": "string", "format": "email", "description": "로그인 식별자. 이메일 형식이어야 하며, 실제 존재 여부 검증 없이 회원 목록 내 중복만 체크한다." },
          "name": { "type": "string", "example": "홍길동" },
          "role": { "$ref": "#/components/schemas/Role" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "required": ["userId", "email", "name", "role", "createdAt", "updatedAt"]
      },
      "SignupRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email", "description": "이메일 형식이어야 하며 기존 회원 목록과 중복될 수 없다." },
          "password": { "type": "string", "format": "password", "minLength": 8 },
          "name": { "type": "string" }
        },
        "required": ["email", "password", "name"]
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "format": "password" }
        },
        "required": ["email", "password"]
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "accessToken": { "type": "string", "description": "JWT Access Token" },
          "user": { "$ref": "#/components/schemas/User" }
        },
        "required": ["accessToken", "user"]
      },
...
    }
  }
}